+++ sharding-core/sharding-core-parse/sharding-core-parse-common/src/main/java/org/apache/shardingsphere/core/parse/sql/segment/dml/pagination/NumberLiteralPaginationValueSegment.java
@@ -29,5 +29,5 @@ public interface NumberLiteralPaginationValueSegment extends PaginationValueSegm
      * 
      * @return value
      */
-    int getValue();
+    long getValue();
 }
